refactor: move data access code from service to repository modules
