+++ apps/grammar-trainer-api/src/app/users/users.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
+import { UsersRepository } from './users.repository';
 
 @Module({
-  providers: [UsersService],
+  providers: [UsersService, UsersRepository],
   exports: [UsersService],
 })
 export class UsersModule {}
