+++ apps/grammar-trainer-api/src/app/users/users.repository.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { UsersRepository } from './users.repository';
+
+describe('UsersRepository', () => {
+  let provider: UsersRepository;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [UsersRepository],
+    }).compile();
+
+    provider = module.get<UsersRepository>(UsersRepository);
+  });
+
+  it('should be defined', () => {
+    expect(provider).toBeDefined();
+  });
+});
